Save parsed moves
Change data layout to save the parsed moves instead of
pgn file offsets.

Now we create 2 files, one with the hashed keys of each position
and one with the parsed moves, every item in the first file points
to the corresponding move in the second one.

In the second file moves of each game are separated by MOVE_NONE.