Sub-sort positions by move frequency
The groups of positions with same key are sorted according to their continuation move score.
Continution moves are first counted and scored according to their frequency.

This is somewhat complicated becuase we don't want to keep a score field in our
position struct, so we need a temporary working struct to do the sorting job.