Huge speedup: don't use std::string in move_is_san()
Allocating a string slows down compared to using
the low level char[]. Normally this is neer the case
but in the super hot paths as is this function taht is
teh most called of all and before this patch was also
the most time consuming

Speed up is huge!

Profiled and found the bottleneck thanks to Kcachegrind
under Ubuntu.