Introduce powerful recovery strategy
For malformed san that miss disambiguation rule,
then play the remaining game for each one of the
possible candidates and if only one survives then
assume this is the intended one.

With this patch we are able to parse "Immortal games":

http://www.angelfire.com/games3/smartbridge/immortal_games.zip

And correctly recover all the ambiguous moves but one.

Note that this strategy works recursively so that we are
able to fix more than one ambiguous move in the same game.