Avoid artifacts in recursive recovering
Previous commit introduced a powerful rule, but to make it correct we should force
strict check (that among other things disables recursivity), then pick the move
that survived longest as our best move, and finally recheck only that move
with relaxed checks.

This scheme, although a bit complex, allow to find the correct move and at the
same time to validate it even if remaining game is still not correct (as is
the case with simple recursion). But differently from recursion, we don't
follow wrong lines.

With this patch immortal_games.pgn is fully and correctly parsed.