Handle spaces in result
Import chessdoctor.pgn but with one wrong game removed.

We had one missing black move there, in game:
{00037.5.22 Analyzed by Candidate Master Daniel Waite}

    16.Rae1 17.Qd2

This is really unrecoverable, so drop the game. There are also another
two fixes for missing braces and for an "e.p." notation in the middle of
the game. I have manually fixed those annotating my changes with "ADDED"
for future reference.

With the above fixes the file is finally parsed with a whopping 249
incorrect moves!