Store games results and ID
Use Polyglot 'learn' parameter to store game result in the upper 2 bits,
and game offset in the PGN file. Note that the offset is 8 bytes aligned
and points to "somewhere" in the game. It is up to the look up tool to
find game's boundaries. This allow us to index up to 8GB PGN files.
Result is stored in the upper 2 bits so that sorting by 'learn'
allows easy counting of result statistics.

When writing a Polyglot book, duplicated positions are removed (this is
how Polyglot standard works). To get game id and result we have to save to
file all the positions, even if duplicated, because they could belong to
different games, and each game with its ID. So to enable this mode, the
tool should be called with 'full' command line argument, like:

     ./parser myfile.pgn full